i think i finished 703s but idk

+++ src/Prog703s/Linux.java
@@ -1,12 +1,12 @@
 package Prog703s;
 
 public class Linux extends Computer{
-  private double myCode;
+  private int myCode;
 
-  public Linux(String name, int num, double val, double code){
+  public Linux(String name, int num, double val, int code){
     super(name, num, val);
     myCode = code;
   }
 
-  public double getCode(){return myCode;}
+  public int getCode(){return myCode;}
 }
